+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 ## Changes
 
+- **Tables / Launch Button**: Added options to launch VPX versions in camera mode.
 - **Tables / Media Previews**: Added tooltips with filename, size and modification date to all previews.
 - **Tables / Media Recorder**: Added new "expert" mode which allows to customize the ffmpeg.exe command for every screen. Please refer to the ffmpeg documentation for more details. And if someone manages to record with audio, please share the corresponding command on my Discord!
 - **Competitions / iScored**: Added error message in case the game room's read API is not enabled.
@@ -22,7 +23,7 @@
   - Fixed auto-renaming for .fpt files.
   - Fixed some validations that have been skipped for .fpt files.
   - Added renaming support for BAM configuration files.
-  - Added upload detection for BAM configuration files, including a separate upload dialog.
+  - Added upload detection for BAM configuration files, including a separate upload dialog. When you upload **Terry Red** bundles, make sure to select only the BAM files you wish to install by using the media filter dialog (which is part of the table upload dialog).
   - Added option in the table deletion dialog for BAM configuration files.